refactor(settings): regroup categories and fully remove enableVibration
- Move key overlay (6 items) from Display to Behavior as the 'Key Overlay'
  group; introduce 'Input' and 'Scroll Send' groups; keep invertPaneNavigation
  flat. Behavior becomes 3 groups + 1 flat = 13 items.
- Move screen-control items (Keep Screen On / Screen Orientation / Max Refresh
  Rate) from Behavior to Display as a new 'Screen' group (10 items).
- Migrate _orientationLabel/_refreshRateLabel + picker imports to display and
  drop them from behavior atomically (R-5/H-1).
- Update search descriptors (category/orderInCategory/groupLabel) and item-count
  comments (Display 10 / Behavior 13 / total 36).
- Fully remove enableVibration from AppSettings/settings_provider (model-layer
  diff limited to enableVibration; migration untouched per user decision Q3).
- Update tests: move KeepScreenOn to display, add Screen Orientation/Refresh
  picker tests, drop Haptic references in behavior/tablet, fix search counts.
