+++ lib/screens/settings/sections/behavior_section.dart
@@ -6,18 +6,16 @@ import '../../../l10n/l10n_ext.dart';
 import '../../../providers/settings_provider.dart';
 import '../../../theme/design_colors.dart';
 import '../../custom_keys/custom_keys_screen.dart';
-import '../pickers/orientation_picker.dart';
-import '../pickers/refresh_rate_picker.dart';
+import '../pickers/overlay_position_picker.dart';
 import '../pickers/scroll_send_input_picker.dart';
 import '../search/settings_search_item.dart';
 import '../settings_category.dart';
+import '../widgets/settings_section_header.dart';
 
-/// Behavior（操作）カテゴリ（フラット・グループ見出しなし）。
+/// Behavior（操作）カテゴリ: キーオーバーレイ / 入力 / スクロール送信 の3グループ + フラット。
 ///
 /// - wheelSendVerifiedProvider / settingsScrollSendUnverifiedNote を同梱（L3/D15）。
 /// - CJK Mode の iOS 判定は `Theme.of(context).platform` を維持（M6/D12）。
-/// - Haptic Feedback（enableVibration）は既知: 現在この設定はハプティクスに
-///   反映されない（A3 注記・別 issue 化推奨）。
 class BehaviorSection extends ConsumerWidget {
   const BehaviorSection({super.key});
 
@@ -29,6 +27,70 @@ class BehaviorSection extends ConsumerWidget {
     return Column(
       crossAxisAlignment: CrossAxisAlignment.stretch,
       children: [
+        SettingsSectionHeader(title: l10n.settingsGroupKeyOverlay),
+        SwitchListTile(
+          secondary: const Icon(Icons.visibility),
+          title: Text(l10n.settingsKeyOverlay),
+          subtitle: Text(l10n.settingsKeyOverlayDescription),
+          value: settings.showKeyOverlay,
+          onChanged: (value) {
+            ref.read(settingsProvider.notifier).setShowKeyOverlay(value);
+          },
+        ),
+        if (settings.showKeyOverlay) ...[
+          SwitchListTile(
+            secondary: const Icon(Icons.keyboard),
+            title: Text(l10n.settingsModifierKeys),
+            subtitle: Text(l10n.settingsModifierKeysDescription),
+            value: settings.keyOverlayModifier,
+            onChanged: (value) {
+              ref.read(settingsProvider.notifier).setKeyOverlayModifier(value);
+            },
+          ),
+          SwitchListTile(
+            secondary: const Icon(Icons.space_bar),
+            title: Text(l10n.settingsSpecialKeys),
+            subtitle: Text(l10n.settingsSpecialKeysDescription),
+            value: settings.keyOverlaySpecial,
+            onChanged: (value) {
+              ref.read(settingsProvider.notifier).setKeyOverlaySpecial(value);
+            },
+          ),
+          SwitchListTile(
+            secondary: const Icon(Icons.arrow_upward),
+            title: Text(l10n.settingsArrowKeys),
+            subtitle: Text(l10n.settingsArrowKeysDescription),
+            value: settings.keyOverlayArrow,
+            onChanged: (value) {
+              ref.read(settingsProvider.notifier).setKeyOverlayArrow(value);
+            },
+          ),
+          SwitchListTile(
+            secondary: const Icon(Icons.shortcut),
+            title: Text(l10n.settingsShortcutKeys),
+            subtitle: Text(l10n.settingsShortcutKeysDescription),
+            value: settings.keyOverlayShortcut,
+            onChanged: (value) {
+              ref.read(settingsProvider.notifier).setKeyOverlayShortcut(value);
+            },
+          ),
+          ListTile(
+            leading: const Icon(Icons.place),
+            title: Text(l10n.settingsOverlayPosition),
+            subtitle: Text(switch (settings.keyOverlayPosition) {
+              'center' => l10n.settingsOverlayPositionCenter,
+              'belowHeader' => l10n.settingsOverlayPositionBelowHeader,
+              _ => l10n.settingsOverlayPositionAboveKeyboard,
+            }),
+            onTap: () => showOverlayPositionPicker(
+              context,
+              ref,
+              settings.keyOverlayPosition,
+            ),
+          ),
+        ],
+        const Divider(),
+        SettingsSectionHeader(title: l10n.settingsGroupInput),
         ListTile(
           leading: const Icon(Icons.apps),
           title: Text(l10n.settingsCustomButtons),
@@ -36,47 +98,6 @@ class BehaviorSection extends ConsumerWidget {
           onTap: () => Navigator.of(
             context,
           ).push(MaterialPageRoute(builder: (_) => const CustomKeysScreen())),
-        ),
-        SwitchListTile(
-          secondary: const Icon(Icons.vibration),
-          title: Text(l10n.settingsHapticFeedback),
-          subtitle: Text(l10n.settingsHapticFeedbackDescription),
-          value: settings.enableVibration,
-          onChanged: (value) {
-            ref.read(settingsProvider.notifier).setEnableVibration(value);
-          },
-        ),
-        SwitchListTile(
-          secondary: const Icon(Icons.brightness_high),
-          title: Text(l10n.settingsKeepScreenOn),
-          subtitle: Text(l10n.settingsKeepScreenOnDescription),
-          value: settings.keepScreenOn,
-          onChanged: (value) {
-            ref.read(settingsProvider.notifier).setKeepScreenOn(value);
-          },
-        ),
-        ListTile(
-          leading: const Icon(Icons.screen_rotation),
-          title: Text(l10n.settingsScreenOrientation),
-          subtitle: Text(_orientationLabel(l10n, settings.screenOrientation)),
-          onTap: () =>
-              showOrientationPicker(context, ref, settings.screenOrientation),
-        ),
-        ListTile(
-          leading: const Icon(Icons.speed),
-          title: Text(l10n.settingsMaxRefreshRate),
-          subtitle: Text(_refreshRateLabel(l10n, settings.refreshRate)),
-          onTap: () =>
-              showRefreshRatePicker(context, ref, settings.refreshRate),
-        ),
-        SwitchListTile(
-          secondary: const Icon(Icons.swipe),
-          title: Text(l10n.settingsInvertPaneNavigation),
-          subtitle: Text(l10n.settingsInvertPaneNavigationDescription),
-          value: settings.invertPaneNavigation,
-          onChanged: (value) {
-            ref.read(settingsProvider.notifier).setInvertPaneNavigation(value);
-          },
         ),
         // CJK Mode: iOSのCJK系IMEでDirectInputが多重送信される場合の
         // 回避手段として旧来（v0.7.0-pre4）の確定送信挙動へ戻す（iOSのみ表示）
@@ -100,6 +121,8 @@ class BehaviorSection extends ConsumerWidget {
             ref.read(settingsProvider.notifier).setKeepKeyboardOnEnter(value);
           },
         ),
+        const Divider(),
+        SettingsSectionHeader(title: l10n.settingsGroupScrollSend),
         // inventory: SETTINGS-UI-INPUT-001
         ListTile(
           leading: const Icon(Icons.mouse),
@@ -156,32 +179,19 @@ class BehaviorSection extends ConsumerWidget {
                 .setAutoFitZoomOnScrollSend(value);
           },
         ),
+        // inventory: SETTINGS-UI-INVERT-001
+        const Divider(),
+        SwitchListTile(
+          secondary: const Icon(Icons.swipe),
+          title: Text(l10n.settingsInvertPaneNavigation),
+          subtitle: Text(l10n.settingsInvertPaneNavigationDescription),
+          value: settings.invertPaneNavigation,
+          onChanged: (value) {
+            ref.read(settingsProvider.notifier).setInvertPaneNavigation(value);
+          },
+        ),
       ],
     );
-  }
-
-  String _orientationLabel(AppLocalizations l10n, String value) {
-    switch (value) {
-      case 'portrait':
-        return l10n.settingsOrientationPortrait;
-      case 'landscape':
-        return l10n.settingsOrientationLandscape;
-      default:
-        return l10n.settingsOrientationAuto;
-    }
-  }
-
-  String _refreshRateLabel(AppLocalizations l10n, String value) {
-    switch (value) {
-      case '120':
-        return '120 Hz';
-      case '90':
-        return '90 Hz';
-      case '60':
-        return '60 Hz';
-      default:
-        return l10n.settingsRefreshRateAuto;
-    }
   }
 
   String _scrollSendInputLabel(AppLocalizations l10n, String value) {
@@ -194,110 +204,135 @@ class BehaviorSection extends ConsumerWidget {
   }
 }
 
-/// Behavior セクションの検索 descriptor（全11項目・フラット）。
+/// Behavior セクションの検索 descriptor（全13項目・3グループ+フラット1）。
 ///
-/// フラットのため groupLabel は null（検索結果 subtitle は所属カテゴリ名表示）。
 /// CJK Mode は iOS 限定表示だが、設定として有効な項目のため検索対象に含める（DR-10 と同趣旨）。
-/// Haptic Feedback（enableVibration）は既知: ハプティクス未反映（A3 注記）。
 final List<SettingsSearchItem> behaviorSearchDescriptors = [
+  // --- キーオーバーレイ（Key Overlay） ---
   SettingsSearchItem(
     category: SettingsCategory.behavior,
     orderInCategory: 0,
-    id: 'customButtons',
-    title: (l10n) => l10n.settingsCustomButtons,
-    description: (l10n) => l10n.settingsCustomButtonsDescription,
-    icon: Icons.apps,
+    id: 'keyOverlay',
+    title: (l10n) => l10n.settingsKeyOverlay,
+    description: (l10n) => l10n.settingsKeyOverlayDescription,
+    groupLabel: (l10n) => l10n.settingsGroupKeyOverlay,
+    icon: Icons.visibility,
   ),
   SettingsSearchItem(
     category: SettingsCategory.behavior,
     orderInCategory: 1,
-    id: 'hapticFeedback',
-    title: (l10n) => l10n.settingsHapticFeedback,
-    description: (l10n) => l10n.settingsHapticFeedbackDescription,
-    icon: Icons.vibration,
-  ),
-  SettingsSearchItem(
-    category: SettingsCategory.behavior,
-    orderInCategory: 2,
-    id: 'keepScreenOn',
-    title: (l10n) => l10n.settingsKeepScreenOn,
-    description: (l10n) => l10n.settingsKeepScreenOnDescription,
-    icon: Icons.brightness_high,
-  ),
-  SettingsSearchItem(
-    category: SettingsCategory.behavior,
-    orderInCategory: 3,
-    id: 'screenOrientation',
-    title: (l10n) => l10n.settingsScreenOrientation,
-    valueLabels: [
-      (l10n) => l10n.settingsOrientationPortrait,
-      (l10n) => l10n.settingsOrientationLandscape,
-      (l10n) => l10n.settingsOrientationAuto,
-    ],
-    icon: Icons.screen_rotation,
-  ),
-  SettingsSearchItem(
-    category: SettingsCategory.behavior,
-    orderInCategory: 4,
-    id: 'maxRefreshRate',
-    title: (l10n) => l10n.settingsMaxRefreshRate,
-    valueLabels: [
-      (l10n) => l10n.settingsRefreshRateAuto,
-      (_) => '120 Hz',
-      (_) => '90 Hz',
-      (_) => '60 Hz',
-    ],
-    icon: Icons.speed,
-  ),
-  SettingsSearchItem(
-    category: SettingsCategory.behavior,
-    orderInCategory: 5,
-    id: 'invertPaneNavigation',
-    title: (l10n) => l10n.settingsInvertPaneNavigation,
-    description: (l10n) => l10n.settingsInvertPaneNavigationDescription,
-    icon: Icons.swipe,
-  ),
-  SettingsSearchItem(
-    category: SettingsCategory.behavior,
-    orderInCategory: 6,
-    id: 'cjkMode',
-    title: (l10n) => l10n.settingsCjkMode,
-    description: (l10n) => l10n.settingsCjkModeDescription,
-    icon: Icons.translate,
-  ),
-  SettingsSearchItem(
-    category: SettingsCategory.behavior,
-    orderInCategory: 7,
-    id: 'keepKeyboardOnEnter',
-    title: (l10n) => l10n.settingsKeepKeyboardOnEnter,
-    description: (l10n) => l10n.settingsKeepKeyboardOnEnterDescription,
+    id: 'modifierKeys',
+    title: (l10n) => l10n.settingsModifierKeys,
+    description: (l10n) => l10n.settingsModifierKeysDescription,
+    groupLabel: (l10n) => l10n.settingsGroupKeyOverlay,
     icon: Icons.keyboard,
   ),
   SettingsSearchItem(
     category: SettingsCategory.behavior,
+    orderInCategory: 2,
+    id: 'specialKeys',
+    title: (l10n) => l10n.settingsSpecialKeys,
+    description: (l10n) => l10n.settingsSpecialKeysDescription,
+    groupLabel: (l10n) => l10n.settingsGroupKeyOverlay,
+    icon: Icons.space_bar,
+  ),
+  SettingsSearchItem(
+    category: SettingsCategory.behavior,
+    orderInCategory: 3,
+    id: 'arrowKeys',
+    title: (l10n) => l10n.settingsArrowKeys,
+    description: (l10n) => l10n.settingsArrowKeysDescription,
+    groupLabel: (l10n) => l10n.settingsGroupKeyOverlay,
+    icon: Icons.arrow_upward,
+  ),
+  SettingsSearchItem(
+    category: SettingsCategory.behavior,
+    orderInCategory: 4,
+    id: 'shortcutKeys',
+    title: (l10n) => l10n.settingsShortcutKeys,
+    description: (l10n) => l10n.settingsShortcutKeysDescription,
+    groupLabel: (l10n) => l10n.settingsGroupKeyOverlay,
+    icon: Icons.shortcut,
+  ),
+  SettingsSearchItem(
+    category: SettingsCategory.behavior,
+    orderInCategory: 5,
+    id: 'overlayPosition',
+    title: (l10n) => l10n.settingsOverlayPosition,
+    valueLabels: [
+      (l10n) => l10n.settingsOverlayPositionAboveKeyboard,
+      (l10n) => l10n.settingsOverlayPositionCenter,
+      (l10n) => l10n.settingsOverlayPositionBelowHeader,
+    ],
+    groupLabel: (l10n) => l10n.settingsGroupKeyOverlay,
+    icon: Icons.place,
+  ),
+  // --- 入力（Input） ---
+  SettingsSearchItem(
+    category: SettingsCategory.behavior,
+    orderInCategory: 6,
+    id: 'customButtons',
+    title: (l10n) => l10n.settingsCustomButtons,
+    description: (l10n) => l10n.settingsCustomButtonsDescription,
+    groupLabel: (l10n) => l10n.settingsGroupInput,
+    icon: Icons.apps,
+  ),
+  SettingsSearchItem(
+    category: SettingsCategory.behavior,
+    orderInCategory: 7,
+    id: 'cjkMode',
+    title: (l10n) => l10n.settingsCjkMode,
+    description: (l10n) => l10n.settingsCjkModeDescription,
+    groupLabel: (l10n) => l10n.settingsGroupInput,
+    icon: Icons.translate,
+  ),
+  SettingsSearchItem(
+    category: SettingsCategory.behavior,
     orderInCategory: 8,
+    id: 'keepKeyboardOnEnter',
+    title: (l10n) => l10n.settingsKeepKeyboardOnEnter,
+    description: (l10n) => l10n.settingsKeepKeyboardOnEnterDescription,
+    groupLabel: (l10n) => l10n.settingsGroupInput,
+    icon: Icons.keyboard,
+  ),
+  // --- スクロール送信（Scroll Send） ---
+  SettingsSearchItem(
+    category: SettingsCategory.behavior,
+    orderInCategory: 9,
     id: 'scrollSendInput',
     title: (l10n) => l10n.settingsScrollSendInput,
     valueLabels: [
       (l10n) => l10n.settingsScrollSendInputWheel,
       (l10n) => l10n.settingsScrollSendInputKey,
     ],
+    groupLabel: (l10n) => l10n.settingsGroupScrollSend,
     icon: Icons.mouse,
   ),
   SettingsSearchItem(
     category: SettingsCategory.behavior,
-    orderInCategory: 9,
+    orderInCategory: 10,
     id: 'invertScrollSendDirection',
     title: (l10n) => l10n.settingsInvertScrollSendDirection,
     description: (l10n) => l10n.settingsInvertScrollSendDirectionDesc,
+    groupLabel: (l10n) => l10n.settingsGroupScrollSend,
     icon: Icons.swipe,
   ),
   SettingsSearchItem(
     category: SettingsCategory.behavior,
-    orderInCategory: 10,
+    orderInCategory: 11,
     id: 'autoFitZoomOnScrollSend',
     title: (l10n) => l10n.settingsAutoFitZoomOnScrollSend,
     description: (l10n) => l10n.settingsAutoFitZoomOnScrollSendDesc,
+    groupLabel: (l10n) => l10n.settingsGroupScrollSend,
     icon: Icons.zoom_out_map,
+  ),
+  // --- フラット（グループなし） ---
+  SettingsSearchItem(
+    category: SettingsCategory.behavior,
+    orderInCategory: 12,
+    id: 'invertPaneNavigation',
+    title: (l10n) => l10n.settingsInvertPaneNavigation,
+    description: (l10n) => l10n.settingsInvertPaneNavigationDescription,
+    icon: Icons.swipe,
   ),
 ];
