+++ test/screens/settings/category_display_test.dart
@@ -1,4 +1,6 @@
+import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
+import 'package:shared_preferences/shared_preferences.dart';
 
 import 'test_helpers.dart';
 
@@ -31,6 +33,66 @@ void main() {
       expect(find.text('System'), findsOneWidget);
       expect(find.text('日本語'), findsOneWidget);
       expect(find.text('English'), findsOneWidget);
+    });
+
+    testWidgets('displays Keep Screen On toggle', (tester) async {
+      await buildSettingsApp(tester);
+      await openCategory(tester, 'Display');
+
+      await scrollUntilFound(tester, find.text('Keep Screen On'));
+      expect(find.text('Keep Screen On'), findsOneWidget);
+    });
+
+    testWidgets('Keep Screen On toggle is interactive', (tester) async {
+      SharedPreferences.setMockInitialValues({});
+      await buildSettingsApp(tester);
+      await openCategory(tester, 'Display');
+
+      await scrollUntilFound(tester, find.text('Keep Screen On'));
+      final tile = find.ancestor(
+        of: find.text('Keep Screen On'),
+        matching: find.byType(SwitchListTile),
+      );
+      expect(tester.widget<SwitchListTile>(tile).value, isTrue); // 既定 ON
+
+      await tester.tap(tile);
+      await tester.pumpAndSettle();
+      expect(tester.widget<SwitchListTile>(tile).value, isFalse);
+    });
+
+    testWidgets('displays Screen Orientation setting and opens picker', (
+      tester,
+    ) async {
+      await buildSettingsApp(tester);
+      await openCategory(tester, 'Display');
+
+      await scrollUntilFound(tester, find.text('Screen Orientation'));
+      expect(find.text('Screen Orientation'), findsOneWidget);
+
+      await tester.tap(find.text('Screen Orientation'));
+      await tester.pumpAndSettle();
+      // 背景の現在値サブタイトルとダイアログ内の両方に現れ得るため findsWidgets
+      expect(find.text('Portrait'), findsWidgets);
+      expect(find.text('Landscape'), findsOneWidget);
+      expect(find.text('Auto (follow device)'), findsWidgets);
+      expect(find.byType(RadioListTile<String>), findsNWidgets(3));
+    });
+
+    testWidgets('displays Max Refresh Rate setting and opens picker', (
+      tester,
+    ) async {
+      await buildSettingsApp(tester);
+      await openCategory(tester, 'Display');
+
+      await scrollUntilFound(tester, find.text('Max Refresh Rate'));
+      expect(find.text('Max Refresh Rate'), findsOneWidget);
+
+      await tester.tap(find.text('Max Refresh Rate'));
+      await tester.pumpAndSettle();
+      // 背景の現在値サブタイトルとダイアログ内の両方に現れ得るため findsWidgets
+      expect(find.text('Auto (highest)'), findsWidgets);
+      expect(find.text('120 Hz'), findsOneWidget);
+      expect(find.byType(RadioListTile<String>), findsNWidgets(4));
     });
   });
 }
